Make VDJ in example page work
