+++ example/src/VDJ/components/Umap.js
@@ -8,7 +8,7 @@ import {
   VerticalLegend,
   useCanvas,
   isValueHighlighted as isHighlighted
-} from 'shahlab-planetarium'
+} from '@shahlab/planetarium'
 
 import Grid from '@material-ui/core/Grid'
 import Slider from '@material-ui/core/Slider'
