+++ package.json
@@ -30,20 +30,7 @@
     "@storybook/addon-essentials": "^6.2.9",
     "@storybook/addon-links": "^6.2.9",
     "@storybook/react": "^6.2.9",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "^8.2.2",
     "cross-env": "^7.0.2",
-    "eslint": "^7.11.0",
-    "eslint-config-prettier": "^6.7.0",
-    "eslint-config-standard": "^14.1.0",
-    "eslint-config-standard-react": "^9.2.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-node": "^11.0.0",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.17.0",
-    "eslint-plugin-standard": "^4.0.1",
-    "lodash": "^4.17.21",
     "microbundle-crl": "^0.13.10",
     "npm-run-all": "^4.1.5",
     "csv-loader": "^3.0.3",
@@ -55,7 +42,6 @@
   "dependencies": {
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
-    "babel-eslint": "^10.0.3",
     "bootstrap": "^4.6.0",
     "cross-env": "^7.0.2",
     "d3": "^5.9.2",
@@ -89,5 +75,8 @@
     "rewire": "^5.0.0",
     "storybook": "^6.2.1",
     "styled-components": "^5.2.3"
+  },
+  "resolutions": {
+    "babel-loader": "8.1.0"
   }
 }
