--- src/styles.module.css
@@ -1,1 +1,0 @@
-/* add css module styles here (optional) */
